Native ANT+ Part 2 - tweaks
Couple of little tweaks to the Native ANT support. Firstly,
data channels such as watts, hr and cadence are now rounded
to an integer (lots of grid lines in ride plot looked nasty).

Secondly, basic support for calibration messages; this has been
limited to SRM messages, but will add support for other devices
as we start to debug them.

Part 3 of this work is still pending, this patch only contains
adjustments to part 2. I suspect there will be further adjustments
as we squash bugs for Quarq and PT users.

+++ src/RealtimeData.cpp
@@ -32,11 +32,11 @@ void RealtimeData::setName(char *name)
 }
 void RealtimeData::setWatts(double watts)
 {
-    this->watts = watts;
+    this->watts = (int)watts;
 }
 void RealtimeData::setHr(double hr)
 {
-    this->hr = hr;
+    this->hr = (int)hr;
 }
 void RealtimeData::setTime(long time)
 {
@@ -52,7 +52,7 @@ void RealtimeData::setWheelRpm(double wheelRpm)
 }
 void RealtimeData::setCadence(double aCadence)
 {
-    cadence = aCadence;
+    cadence = (int)aCadence;
 }
 void RealtimeData::setLoad(double load)
 {
